restructure metrics info
